feat: split e2e tests into independently-runnable lanes
Refactor the monolithic e2e_test.go into three labeled test lanes
(operator, bootc, auth) that run in isolated namespaces and can be
triggered independently via Makefile targets or CI comment commands.
